REFACTOR: SecurityUtil을 사용하여 userId 추출 로직 최적화

+++ src/main/kotlin/syntax/backend/runways/dto/ResponseCommentDTO.kt
@@ -10,7 +10,7 @@ data class ResponseCommentDTO (
     val updatedAt : LocalDateTime,
     val author : String,
     val parent : UUID?,
-    val childCount : Long,
+    val childCount : Int,
     val imageUrl : String?,
     val maker : Boolean
 )
